scripts/mount: tmpfs source as upperdir
on newer devices with native casefolding via f2fs, mounting the module with
f2fs as source fucks up due to ovl_dentry_weird.

This is a workaround to that.
What this does is, copies the module to tmpfs, mirrors all selinux context
and then mounts it with tmpfs as the source.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>